+++ CHANGES.md
@@ -11,6 +11,8 @@ unreleased
     - Add a test to ensure the code fragment exhibited in issue #1118 no longer makes Merlin crash (#1996)
     - Add a test case illustrating how a snippet produces two unrelated errors in issue #2000. (#2003)
     - Add a test reproducing issue #1983 where `document` command which sometime concatenates consecutive variants and labels (#2005)
+    - Signature-help should trigger on unfinished `let ... in` bindings (#2009)
+
 
 merlin 5.6
 ==========